Update Executable.java
Updated: Type E is specified to String when class is instantiated. (So don't need to cast (E)) Every E appeared is changed to String.
Updated: local variables names "from" and "to" in "btncontinue.addActionListener" are changed to "sourceOfGraph" and “sinkOfGraph” to distinguish "from" and "to" used somewhere else to represent endpoints of a normal edge.